# swagger/swagger.json
{"swagger":"2.0","info":{"title":"The Cassandra-Wannabe API","description":"CRUD API for a distributed database.","license":{"name":"MIT","url":"https://github.com/freddygv/cassandra-wannabe/blob/master/LICENSE"},"version":""},"host":"localhost:8080","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/db/v1/":{"put":{"tags":["rating"],"summary":"upsert rating","description":"Adds a rating record","operationId":"rating#upsert","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/RatingPayload"}}],"responses":{"204":{"description":"No Content"}},"schemes":["http"]}},"/db/v1/status":{"get":{"tags":["health"],"summary":"health health","description":"Status of the API.","operationId":"health#health","produces":["text/plain"],"responses":{"200":{"description":"OK"}},"schemes":["http"]}},"/db/v1/{movieId}/{userId}":{"get":{"tags":["rating"],"summary":"read rating","description":"Retrieves a rating record","operationId":"rating#read","produces":["application/cassandra.wannabe.rating+json"],"parameters":[{"name":"movieId","in":"path","required":true,"type":"integer"},{"name":"userId","in":"path","required":true,"type":"integer"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/rating"}}},"schemes":["http"]},"delete":{"tags":["rating"],"summary":"delete rating","description":"Delete rating record","operationId":"rating#delete","parameters":[{"name":"movieId","in":"path","required":true,"type":"integer"},{"name":"userId","in":"path","required":true,"type":"integer"}],"responses":{"202":{"description":"Accepted"}},"schemes":["http"]}},"/swagger.json":{"get":{"summary":"Download public/swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"RatingPayload":{"title":"RatingPayload","type":"object","properties":{"movieId":{"type":"integer","example":138493,"minimum":1},"rating":{"type":"number","example":3.5,"minimum":0.5,"maximum":5},"userId":{"type":"integer","example":56757,"minimum":1}},"example":{"movieId":138493,"rating":3.5,"userId":56757},"required":["movieId","userId","rating"]},"rating":{"title":"Mediatype identifier: application/cassandra.wannabe.rating+json; view=default","type":"object","properties":{"movieId":{"type":"integer","example":138493,"minimum":1},"rating":{"type":"number","example":3.5,"minimum":0.5,"maximum":5},"userId":{"type":"integer","example":56757,"minimum":1}},"description":"A movie rating by a user (default view)","example":{"movieId":138493,"rating":3.5,"userId":56757},"required":["movieId","userId","rating"]}},"responses":{"Accepted":{"description":"Accepted"},"NoContent":{"description":"No Content"}}}